feat: add tool use confirmation support
