+++ src/utils/runAgent.ts
@@ -1,6 +1,7 @@
 import { query, type SDKUserMessage } from "@anthropic-ai/claude-agent-sdk"
 import type { AgentChatConfig } from "store"
 import { buildSystemPrompt } from "utils/getPrompt"
+import { createCanUseTool } from "utils/canUseTool"
 
 export const messageTypes = {
   ASSISTANT: "assistant",
@@ -65,43 +66,10 @@ export const createAgentQuery = (options: CreateAgentQueryOptions) => {
           }
         : undefined,
       canUseTool: onToolPermissionRequest
-        ? async (toolName: string, input: any) => {
-            // Notify UI about the permission request
-            onToolPermissionRequest(toolName, input)
-
-            // Wait for user decision via messageQueue
-            const userResponse = await new Promise<string>((resolve) => {
-              messageQueue.push({ resolve })
-            })
-
-            const response = userResponse.toLowerCase().trim()
-
-            // Enter pressed (empty) or explicit "y"/"yes"/"allow"
-            if (!response || ["y", "yes", "allow"].includes(response)) {
-              return { behavior: "allow" as const, updatedInput: input }
-            }
-
-            // ESC or explicit "n"/"no"/"deny"
-            if (["n", "no", "deny"].includes(response)) {
-              return {
-                behavior: "deny" as const,
-                message: "User denied permission",
-                interrupt: true,
-              }
-            }
-
-            // Any other input = modified input
-            try {
-              const updatedInput = JSON.parse(response)
-              return { behavior: "allow" as const, updatedInput }
-            } catch {
-              // If not valid JSON, treat as text modification
-              return {
-                behavior: "allow" as const,
-                updatedInput: { value: response },
-              }
-            }
-          }
+        ? createCanUseTool({
+            messageQueue,
+            onToolPermissionRequest,
+          })
         : undefined,
     },
   })
